Add comments and refactor code - Rename patch file - Add patch entry
