+++ frappe/patches.txt
@@ -218,3 +218,4 @@ frappe.patches.v11_0.update_list_user_settings
 frappe.patches.v11_0.rename_workflow_action_to_workflow_action_master #13-06-2018
 frappe.patches.v11_0.rename_email_alert_to_notification #13-06-2018
 frappe.patches.v11_0.delete_duplicate_user_permissions
+frappe.patches.v11_0.replicate_old_user_permissions
